ResumeController finished

+++ src/main/java/com/softeams/poSystem/core/dtos/ResumeVentasDto.java
@@ -3,7 +3,7 @@ package com.softeams.poSystem.core.dtos;
 import java.math.BigDecimal;
 
 public record ResumeVentasDto(
-        Integer totalVentas,
+        Long totalVentas,
         BigDecimal totalIngresos
 ) {
 }
